test(lang-v2): cover realloc payer target guard
